// src/app/core/models/carDto.ts
export interface CarDto {
    id?: number;
    brandCarId: number;
    brandCarString?: string;
    reference: string;
    price: number;
    modelYear: number;
    color: string;
    horsepower: number;
    numberDoor: number;
    engine: number;
    transmission: string;
    fuelType: string;
    traction: number;
    numberSeat: number;
    steering: string;
    category: string;
    stock: number;
    imagePath: string;
  }